+++ voting/frontend/app/layout.tsx
@@ -1,12 +1,20 @@
 import './globals.css';
 import '@rainbow-me/rainbowkit/styles.css';
-import {Providers} from './providers';
+
+import { Providers } from './providers';
+import Headers from "@/components/headers/headers";
 
 function RootLayout({ children }: { children: React.ReactNode }) {
     return (
         <html lang="en">
             <body className="bg-gray-50">
-                <Providers>{children}</Providers>
+                <Providers>
+                    <Headers />
+
+                    <main className='container mx-auto py-10 px-5'>
+                        {children}
+                    </main>
+                </Providers>
             </body>
         </html>
     );