+++ voting/frontend/app/page.tsx
@@ -1,11 +1,9 @@
 "use client";
 
-import Headers from "@/components/headers/headers";
-
 export default function Home() {
   return (
         <div >
-            <Headers />
+            <h1>Home</h1>
         </div>
   )
 }